fixed routes in actions, renamed files, cleaned up for new backend.
