+++ src/components/Dashboard/OldReviewCard.js
@@ -3,16 +3,18 @@ import React, { useState, useRef } from 'react';
 import { connect } from 'react-redux';
 import ReactGA from 'react-ga';
 // actions
-import getCompanyReview from '../../state/actions/index';
-import deleteCompanyReview from '../../state/actions/index';
+import getReview from '../../state/actions/index';
+import deleteReview from '../../state/actions/index';
 
 // icons
-import { TiCalendar } from 'react-icons/ti';
-import { FiThumbsUp } from 'react-icons/fi';
-import { MdPerson } from 'react-icons/md';
-import { FaDollarSign, FaRegClock, FaRegMoneyBillAlt } from 'react-icons/fa';
-import { GoLocation } from 'react-icons/go';
-
+import {
+  TiLocationOutline,
+  TiArchive,
+  TiThumbsUp,
+  TiThumbsDown,
+  TiGlobe
+} from 'react-icons/ti';
+import { GiWeightLiftingUp } from 'react-icons/gi';
 // styles
 import {
   Box,
@@ -25,7 +27,6 @@ import {
   ModalCloseButton,
   Button,
   Icon,
-  Badge,
   PseudoBox,
   AlertDialog,
   AlertDialogBody,
@@ -37,14 +38,12 @@ import {
   useDisclosure
 } from '@chakra-ui/core';
 
-const ReviewCard = ({
-  review,
-  reviewDeleted,
-  history,
-  deleteCompanyReview
-}) => {
+const ReviewCard = ({ review, reviewDeleted, history, deleteReview }) => {
   //allows the use of toasts
   const toast = useToast();
+
+  //toggle to determine whether deleting worked
+  const [deleteSuccess, setDeleteSuccess] = useState(false);
 
   // basic usage for the SingleReview modal
   const { isOpen, onOpen, onClose } = useDisclosure();
@@ -57,12 +56,12 @@ const ReviewCard = ({
 
   //routes to single review
   const navToEditRoute = () => {
-    history.push(`/dashboard/${review.company_review_id}`);
+    history.push(`/dashboard/${review.interview_review_id}`);
   };
 
   //deletes the review in question
   const submitDelete = () => {
-    deleteCompanyReview(review.user_id, review.company_review_id).then(() => {
+    deleteReview(review.user_id, review.interview_review_id).then(() => {
       window.location.reload();
       // history.push('/dashboard')
       toast({
@@ -105,164 +104,148 @@ const ReviewCard = ({
       {/* ------------------------------------------------------------------------------------------------ */}
       <Modal isOpen={isOpen} onClose={onClose} size='80%'>
         <ModalOverlay />
-        <ModalContent w='100%' py='3%' wrap='nowrap'>
+        <ModalContent w='100%' px='12%' py='3%' wrap='nowrap'>
           <ModalCloseButton background='none' border='none' />
 
           {/* Basic info container */}
-          <Flex flexDir={{ lg: 'row', sm: 'column' }} align='center' mx='8%'>
+          {/* <SingleReview /> */}
+          <Flex align='center'>
             <Flex align='center'>
               <Avatar size='2xl' src={`//logo.clearbit.com/${review.domain}`} />
             </Flex>
-            <Flex flexDir='column' pl={{ lg: '8%', sm: '0%' }} width='100%'>
+            <Flex flexDir='column' pl='4%'>
+              <Flex as='h1' w='100%' align='center' wrap='nowrap'>
+                {review.company_name} Interview Review
+              </Flex>
               <Flex
                 as='h2'
+                fontSize='lg'
                 w='100%'
+                fontWeight='medium'
                 align='center'
-                justify={{ lg: 'flex-start', sm: 'center' }}
                 wrap='nowrap'
               >
-                {review.company_name}
+                <Flex>Interview Review:</Flex>
+                <Flex ml='2%'>
+                  {Array(5)
+                    .fill('')
+                    .map((_, i) => (
+                      <Icon
+                        name='star'
+                        key={i}
+                        color={
+                          i < review.interview_rating ? 'black' : 'gray.300'
+                        }
+                      />
+                    ))}
+                </Flex>
               </Flex>
-              <Flex justify='space-between'>
-                <Flex flexDir='column'>
-                  <Flex fontSize='small' fontWeight='light' color='#9194A8'>
-                    Location
-                  </Flex>
-                  <Flex>Mountain View, CA</Flex>
-                </Flex>
-                <Flex flexDir='column'>
-                  <Flex fontSize='small' fontWeight='light' color='#9194A8'>
-                    Job Title
-                  </Flex>
-                  <Flex>{review.job_title}</Flex>
-                </Flex>
-                <Flex flexDir='column'>
-                  <Flex fontSize='small' fontWeight='light' color='#9194A8'>
-                    Rating
-                  </Flex>
-                  <Flex>
-                    Company Rating
-                    <Flex>
-                      {Array(5)
-                        .fill('')
-                        .map((_, i) => (
-                          <Icon
-                            name='star'
-                            key={i}
-                            color={i < review.job_rating ? 'black' : 'gray.300'}
-                            ml='5px'
-                          />
-                        ))}
-                    </Flex>
-                  </Flex>
-                </Flex>
+              <Flex as='p' fontSize='md' align='center' wrap='nowrap'>
+                Position: {review.job_title}
               </Flex>
             </Flex>
           </Flex>
 
           {/* Secondary info container */}
-          <Flex w='100%' backgroundColor='#344CD0' color='white' mt='4%'>
-            <Flex
-              w='100%'
-              overflow='hidden'
-              justify='space-evenly'
-              align='center'
-              py='1%'
-            >
+          <Flex w='100%' flexDir='column'>
+            <Flex w='100%' mt='1.5%' overflow='hidden' justify='space-between'>
               <Flex align='center' wrap='nowrap'>
-                <Box as={FaDollarSign} size='2em' mr='5px'></Box>
-                <Flex flexDir='column'>
-                  <Flex as='h3' fontWeight='light' fontSize='md' isTruncated>
-                    ${review.salary}
-                  </Flex>
-                  <Flex as='h3' fontWeight='light' fontSize='sm' isTruncated>
-                    Salary
-                  </Flex>
+                <Box as={TiLocationOutline} mr='5px'></Box>
+                <Flex as='h3' fontWeight='light' fontSize='lg' isTruncated>
+                  {review.job_location}
                 </Flex>
               </Flex>
               <Flex align='center' wrap='nowrap'>
-                <Box as={FiThumbsUp} size='2em' mr='5px'></Box>
-                <Flex flexDir='column'>
-                  <Flex as='h3' fontWeight='light' fontSize='md' isTruncated>
-                    {review.typical_hours} hrs week
-                  </Flex>
-                  <Flex as='h3' fontWeight='light' fontSize='sm' isTruncated>
-                    Working Hours
-                  </Flex>
+                {review.offer_received ? (
+                  <>
+                    <Box as={TiThumbsUp} mr='5px'></Box>
+                    <Flex as='h3' fontWeight='light' fontSize='lg'>
+                      Received 0ffer
+                    </Flex>{' '}
+                  </>
+                ) : (
+                  <>
+                    {' '}
+                    <Box as={TiThumbsDown} mr='5px'></Box>
+                    <Flex as='h3' fontWeight='light' fontSize='lg'>
+                      No Offer
+                    </Flex>{' '}
+                  </>
+                )}
+              </Flex>
+              <Flex align='center' wrap='nowrap'>
+                <Box as={GiWeightLiftingUp} mr='5px'></Box>
+                <Flex as='h3' fontWeight='light' fontSize='lg' isTruncated>
+                  Difficult
                 </Flex>
               </Flex>
               <Flex align='center' wrap='nowrap'>
-                <Box as={MdPerson} size='2em' mr='5px'></Box>
-                <Flex flexDir='column'>
-                  <Flex as='h3' fontWeight='light' fontSize='md' isTruncated>
-                    {review.work_status}
-                  </Flex>
-                  <Flex as='h3' fontWeight='light' fontSize='sm' isTruncated>
-                    Status
-                  </Flex>
-                </Flex>
-              </Flex>
-              <Flex align='center' wrap='nowrap'>
-                <Box as={TiCalendar} size='2em' mr='5px'></Box>
-                <Flex flexDir='column'>
-                  <Flex as='h3' fontWeight='light' fontSize='md' isTruncated>
-                    {review.start_date} - {review.end_date}
-                  </Flex>
-                  <Flex as='h3' fontWeight='light' fontSize='sm' isTruncated>
-                    Date
-                  </Flex>
+                <Box as={TiGlobe} mr='5px'></Box>
+                <Flex as='h3' fontWeight='light' fontSize='lg' isTruncated>
+                  4 Rounds
                 </Flex>
               </Flex>
             </Flex>
           </Flex>
 
-          {/* Topics container */}
+          {/* Types container */}
           <Flex
             as='h2'
             fontWeight='medium'
             fontSize='xl'
             w='100%'
-            mt='2%'
+            mt='3%'
             mb='1.5%'
-            px='8%'
             overflow='hidden'
           >
-            Topics
+            Interview Types
           </Flex>
-          <Flex
-            justify='space-between'
-            wrap='wrap'
-            whiteSpace='nowrap'
-            mb='2%'
-            px='8%'
-          >
+          <Flex justify='space-between' wrap='wrap' whiteSpace='nowrap' mb='2%'>
             <Flex as='p' bg='#F2F6FE' px='1%' mb='1.5%'>
-              Career Growth
+              Phone screening
             </Flex>
             <Flex as='p' bg='#F2F6FE' px='1%' mb='1.5%'>
-              Benefits
+              Online coding test
             </Flex>
             <Flex as='p' bg='#F2F6FE' px='1%' mb='1.5%'>
-              Salary
+              Behavioral questions
             </Flex>
             <Flex as='p' bg='#F2F6FE' px='1%' mb='1.5%'>
-              Company Culture
+              Case interviews
             </Flex>
             <Flex as='p' bg='#F2F6FE' px='1%' mb='1.5%'>
-              Another Cool Thing
+              Whiteboarding
             </Flex>
           </Flex>
 
           {/* Review container */}
-          <Flex
-            as='p'
-            w='100%'
-            wrap='nowrap'
-            overflow='hidden'
-            px='8%'
-            align='center'
-          >
-            {review.comment}
+          <Flex as='p' w='100%' wrap='nowrap' overflow='hidden'>
+            {review.job_review}
+          </Flex>
+
+          <Flex w='100%' mt='3%' align='center'>
+            <Flex fontWeight='medium' fontSize='xl' wrap='nowrap'>
+              Overall Rating
+            </Flex>
+            <Flex ml='1.5%'>
+              {Array(5)
+                .fill('')
+                .map((_, i) => (
+                  <Icon
+                    name='star'
+                    key={i}
+                    color={i < review.job_rating ? 'black' : 'gray.300'}
+                  />
+                ))}
+            </Flex>
+          </Flex>
+          <Flex as='h2' w='100%' wrap='nowrap' overflow='hidden' mt='1.5%'>
+            <Flex fontWeight='medium' fontSize='xl'>
+              Salary Offered
+            </Flex>
+            <Flex ml='1.5%' fontWeight='light' fontSize='xl'>
+              ${review.salary}
+            </Flex>
           </Flex>
 
           <ModalFooter>
@@ -320,51 +303,29 @@ const ReviewCard = ({
           </ModalFooter>
         </ModalContent>
       </Modal>
-
       {/* ------------------------------------------------------------------------------------------------ */}
       {/* ---------------------------------------DashBoard Cards------------------------------------------ */}
       {/* ------------------------------------------------------------------------------------------------ */}
-
       {/* Review container */}
       <PseudoBox
-        width={[
-          '100%', // base
-          '100%', // 480px upwards
-          '100%', // 768px upwards
-          '45%' // 992px upwards
-        ]}
+        w='45%'
+        // h='50%'
         mt='3%'
         mx='2.5%'
-        px='4%'
-        py='2%'
-        background='#FBFCFF'
+        p='3.5%'
+        wrap='nowrap'
+        background='#F2F6FE'
         borderRadius='12px'
         display='flex'
-        flexDir='column'
         justifyContent='center'
         alignItems='center'
         _hover={{ bg: '#4EADF9', color: 'white' }}
         onClick={onOpen}
       >
         {/* Review content container */}
-        <Flex width='100%' justifyContent='flex-end'>
-          <Badge
-            backgroundColor='#344CD0'
-            color='white'
-            fontSize='1em'
-            fontWeight='light'
-            rounded='full'
-            px='15px'
-            overflow='hidden'
-          >
-            Web
-          </Badge>
-        </Flex>
-
-        {/* Review content container */}
-        <Flex flexDir='column'>
+        <Flex w='100%' wrap='wrap' justify='right' alignContent='center'>
           {/* headline line container  */}
-          <Flex w='100%' h='100px'>
+          <Flex w='100%' h='100px' mb='3%'>
             {/* avatar box */}
             <Box justify='center' align='center' h='88px' mr='36px'>
               <Avatar size='xl' src={`//logo.clearbit.com/${review.domain}`} />
@@ -375,64 +336,94 @@ const ReviewCard = ({
                 as='h2'
                 w='100%'
                 align='center'
+                wrap='wrap'
                 overflow='hidden'
                 isTruncated
               >
-                {review.company_name} company review
+                {review.company_name} interview review
               </Flex>
               <Flex width='100%'>
-                <Flex as='h4' align='center'>
-                  {review.job_rating}.0
+                <Flex as='h4' align='center' wrap='nowrap'>
+                  {review.overall_rating}.0
                 </Flex>
-                <Flex align='center'>
+                <Flex align='center' wrap='nowrap'>
                   {Array(5)
                     .fill('')
                     .map((_, i) => (
                       <Icon
                         name='star'
                         key={i}
-                        color={i < review.job_rating ? '#344CD0' : 'gray.300'}
+                        color={
+                          i < review.overall_rating ? '#344CD0' : 'gray.300'
+                        }
                         ml='8%'
                       />
                     ))}
                 </Flex>
               </Flex>
+              {/* <Flex align='center' w='40%' wrap='nowrap'> */}
               <Flex as='p' w='100%' fontWeight='light'>
                 Position: {review.job_title}
               </Flex>
             </Flex>
+            {/* </Flex> */}
           </Flex>
 
-          {/* Second main container */}
+          {/* Company name & location container */}
           <Flex
+            justify='flex-start'
             w='100%'
             font-size='18'
             fontWeight='light'
             justify='space-evenly'
             align='center'
+            wrap='nowrap'
             mb='1%'
+            mt=''
           >
-            <Flex align='center'>
-              <Box as={FaRegMoneyBillAlt} mr='10px'></Box>
+            <Flex align='center' wrap='nowrap'>
+              <Box as={TiArchive} mr='10px'></Box>
               <Flex as='p' overflow='hidden'>
                 ${review.salary}
               </Flex>
             </Flex>
-            <Flex align='center'>
-              <Box as={GoLocation} mr='10px'></Box>
-              <Flex as='p'>MISSING LOCATION</Flex>
-            </Flex>
-            <Flex align='center'>
-              <Box as={FaRegClock} mr='10px'></Box>
+            <Flex align='center' wrap='nowrap'>
+              <Box as={TiLocationOutline} mr='10px'></Box>
               <Flex as='p'>
-                {review.start_date}-{review.end_date}
+                {review.city}, {review.abbreviation}
               </Flex>
             </Flex>
+            <Flex align='center' wrap='nowrap'>
+              {review.offer_status ? (
+                <>
+                  <Box as={TiThumbsUp} mr='10px'></Box>
+                  <Flex as='p'>Received Offer</Flex>{' '}
+                </>
+              ) : (
+                <>
+                  {' '}
+                  <Box as={TiThumbsDown} mr='10px'></Box>
+                  <Flex as='p' mr='10px'>
+                    No Offer
+                  </Flex>{' '}
+                </>
+              )}
+            </Flex>
           </Flex>
-
+          {/* headline line container  */}
+          <Flex w='100%' align='center' wrap='nowrap' pl='1%' mt='1%'>
+            <h2>{review.tagline}</h2>
+          </Flex>
           {/* summary container */}
-          <Flex w='100%' h='95px' overflow='hidden'>
-            <p>{review.comment}</p>
+          <Flex
+            w='100%'
+            h='50%'
+            wrap='nowrap'
+            overflow='hidden'
+            pl='1%'
+            mt='0.5%'
+          >
+            <p>{review.job_review}</p>
           </Flex>
         </Flex>
       </PseudoBox>
@@ -445,4 +436,4 @@ const mapStateToProps = state => {
     reviewDeleted: state.review.reviewDeleted
   };
 };
-export default connect(mapStateToProps, deleteCompanyReview)(ReviewCard);
+export default connect(mapStateToProps, (getReview, deleteReview))(ReviewCard);
