+++ src/state/types/index.js
@@ -49,24 +49,3 @@ export const FETCH_COMPANIES_FAILURE = 'FETCH_COMPANIES_FAILURE';
 export const POST_COMPANY_START = 'POST_COMPANY_START';
 export const POST_COMPANY_SUCCESS = 'POST_COMPANY_SUCCESS';
 export const POST_COMPANY_FAILURE = 'POST_COMPANY_FAILURE';
-
-// ============ FETCH COMPANY REVIEW LIST ===========
-
-export const FETCH_COMPANY_REVIEWS_START = 'FETCH_COMPANY_REVIEWS_START';
-export const FETCH_COMPANY_REVIEWS_SUCCESS = 'POST_COMPANY_REVIEW_SUCCESS';
-export const FETCH_COMPANY_REVIEWS_FAILURE = 'FETCH_COMPANY_REVIEWS_FAILURE';
-
-// ============ FETCH COMPANY REVIEW BY ID ===========
-
-export const FETCH_COMPANY_REVIEW_BY_ID_START =
-	'FETCH_COMPANY_REVIEW_BY_ID_START';
-export const FETCH_COMPANY_REVIEW_BY_ID_SUCCESS =
-	'FETCH_COMPANY_REVIEW_BY_ID_SUCCESS';
-export const FETCH_COMPANY_REVIEW_BY_ID_FAILURE =
-	'FETCH_COMPANY_REVIEW_BY_ID_FAILURE';
-
-// ============ POST COMPANY REVIEW ===========
-
-export const POST_COMPANY_REVIEW_START = 'POST_COMPANY_REVIEW_START';
-export const POST_COMPANY_REVIEW_SUCCESS = 'POST_COMPANY_REVIEW_SUCCESS';
-export const POST_COMPANY_REVIEW_FAILURE = 'POST_COMPANY_REVIEW_FAILURE';
